一级回复已完成未测试

+++ src/main/java/com/community/model/Comment.java
@@ -13,7 +13,7 @@ public class Comment {
     private Long id;
     private Long parentId;
     private Integer type;
-    private Integer commentator;
+    private Long commentator;
     private Long gmtCreate;
     private Long gmtModified;
     private Long likeCount;
